clean up allTags
Move allTags to a collection rather than a filter
